Refactor ListItem and CommentsList components
